task003: разнесена логика по соответствующим файлам
